changes add_targets, debugged use_keras
changes of add_targets need to be verified. I am of the opinioin that the smoothing part should never been called but I saw a difference after introdcuing it. The cleanup of lossix / winix incremented by 1 and not by slot.
Changes need to be verified with unit tests and visualization.
use_keras bugs removed.